Upgrade to aiida-workgraph v0.4.10
To include several bugfixes we upgrade to the version v0.4.10. This is
refuired to specify computer in the metadata of the workgraph task as
this is not possible with the currently pinned version. The workgraph
API has changed therefore we adapt to new API the functions which makes
part of the function calls cleaner. The workaround is still needed but
has been better documented.