Moves logic from `CanonicalWorkflow` to `core.Workflow` and `ConfigWorkflow`
The validation of `cycles` and `tasks` in `list_not_empty` is moved to
the `ConfigWorkflow`. This subsequently required a modification of the
tests that initalize a `ConfigWorkflow` with empty `cycles` and `tasks`
as this is not anymore allowed through the tests.

The `data_dict` and `task_dict` member variables are moved to the
constructor of the `core.Workflow`. The constructor of `core.Workflow`
is split into two: `from_config_workflow` constructor, which replicates
the behavir of the previous constructor and the default construcor now
accepts all required that are extracted from the passed `ConfigWorkflow`
in the previous constructor as individual parameters.

The introduction of the `from_config_file` constructor in
ConfigWorkflow, which replicates the behavior of `load_workflow_config`,
allows the `rootdir` to become a part of `ConfigWorkflow`. This approach
eliminates the need for an external utility function that essentially
acts as a constructor for `ConfigWorkflow`, simplifying the interface by
enabling direct access to the `rootdir` within `core.Workflow`.