Fixing default argument only if rootdir is not specified and adding name option
In the last commit, the implementation did not allow specifying the
rootdir in the YAML file, as it would be automatically overwritten in
the from_config_file constructor. To address this, we updated the
process to explicitly parse the YAML file into a generic Python object
of nested dictionaries, followed by its conversion into a validated
Pydantic object using the parse_yaml_file_as function. This approach
enables us to directly work with the parsed object and determine whether
the user provided a value for name and rootdir in the yaml file.
Consequently, this removes the prior requirement of enforcing a default
value for rootdir, which was introduced in the last commit for technical
simplicity.

After looking how pydantic-yaml parses files, it became apparent that
they still use the deprecated version to load objects with v1 API. Since
we can easily replace this functionality, there is no need to keep the
dependency. But this should happen in another PR.