Refactor namelists of icon namelist
Refactor of `Namelist` and ¨IconTask`:
- Reduce `IconTask` API to `update_namelists_from_workflow` as other
  functions could be moved to other places or were note needed .
- Separate dumping of namelists from the update process as one can pass
  the namelist by a memory buffer (io.StringIO) in AiiDA.
- Be more consistent with the `Config<CLASS_NAME>Spec` naming convention
  and usage. This class is used to for shared member variables between
  the config file and core representation.

Convert all path variables in config schema to `pathlib.Path`.

Introduce private variables into `IconTask` and `Namelist` for stricter
encapsulationb by using this design

```python
from dataclasses import InitVar, dataclass

@dataclass
class A:
  variable: InitVar[int]
  def __post_init__(self, variable: int):
    self._variable = variable

  @property
  def variable(self) -> int:
    return self._variable

a = A()
print(a.variable) # ok
a.variable = 5 # does not work
```
This pattern however clashes with mypy type checking. It erroneously
detects an overwrite in the definition of `@property`. With `InitiVar`
the variable is however only available during initialization.