Add integration tests with icon to CI
New CI workflow icon-tests that installs icon using spack and runs the
tests assigned with the marker icon_required.

We use the spack buildcache feature to reuse a previous icon installation in
the CI to reduce the CI time drastically (from 70 mins to 3 mins)

Because we need to activate the spack shell to access the icon installation
the python version that required for icon is also activated. We thus use for
this new workflow the spack python version instead of the one of
actions/setup-python like in the unittests workflow.

Introduce `requires_icon` marker and add test that checks if icon can be
found.