Support for copy of local scripts to remote machine
Support the copy of local scripts to remote machine by adding the script
specified in `src` to the nodes of the `ShellJob`. Because aiida does
not copy the file over preserving the execution permission, we cannot
support a syntax of the form `./icon.py`. Therefore I changed the usage
to the syntax "python icon.py".

Introduce localhost enum to only expand env variables when localhost. I
think this will be removed as the support of environment variables is
quite fragile since it opens the door for multiple valid uses cases that
are hard to support at the same time.