Move expansion of environmental variables to tests.
The support of environment variables in the config yaml makes it unclear when
these variables are expanded. For local files that should be moved to remote,
the expansion should be before runtime, while for files that are on the remote
it should happen during runtime. Since we do not have a clear local keyword for
files, this feature needs more consideration is for now removed.

For the tests, the environment variables are however required as we need to be
able to specify relative paths, since the test directory changes for different
dev. Before this PR we expanded the environment variables in the code. Since
this mechanic is removed by this PR, we perform the expansion now in the pytest
fixtures before passing the file to sirocco Because this needs a modification
of the config files, to not touch the original file we copy the test case
folder to a tmp path.

Now we added back the checks that the path are absolute. Before this PR
this could not be done, since it had to be considered that an
environment variable makes the path relative. The check for
`ShellTask.src` is skipped because it will be enforced to be relative in
PR #172.