Rename `src` to `path` and `relpath` and include all paths to repr
Rename `src` to `relpath` if it is required to be a relative path (in
case of `NamelistFile` and `ShellTask`).  Rename `src` to `path` if it
is required to be an absolute path (in case of `AvailableData`). By this
change we can include the `relpath` into the representation. We also
include `IconTask.bin` in the representation as this has been made
absolute in PR #153 (commit 8b5ec4da) and thus does not create a path
that is dependendent on the local development environment.