Enable tests for py310, py311 and p312 in hatch options
Due to the hatch setup the tests in the CI only ran on py312. It seems
that hatch python discoverability detected a py312 version on the runner
and therefore always defaulted to py312.

It is needed to add to the python version to the test matrix to support
the other python versions for running with hatch test.