Add `wrapper_script` field to `IconTask` (#210)
Add to yaml file a field for the `wrapper_script` that is added to the
submission script before the code. This is added to the `CalcJob` wrapper_script
parameter in `IconCalculation`. If no `wrapper_script` is specified, then a
script defined in `aiida-icon` is used that has the defaults for CSCS machines.

Further the name of the aiida icon task is added so it can be accessed in the
test with the label we give it based on the task type and date-timestamp.